We have init!
We check both FLASH0:init.toml as well as SD0:init.toml for potential
programs.

The programs will start, and are optionally kept running.

for example:

[[apps]]
name = "framebuffer_test"
path = "FLASH0:framebuffer_test.elf"
restart_on_failure = false
stack_size = 8192
args = []

will start the framebuffer test, one time, if it gets killed it will not
be started again.

Additionally there is a "run_once" parameter, when set to "true" the
program will be started once. The successful start is recorded in NVS
and a program by that "name" field will not be started on subsequent
boots.

Note that "run_once" is true, and "restart_on_failure" is a valid
configuration. It will just only happen once, at the next boot.