Fix tearing on single-buffered windows
By doing something that is a bit more in line with modern graphics APIs
we now use the P4 MMU to very quickly swap the actual pointers of the
framebuffers for double buffered windows.

This means that APIs that expect to "hand off" their framebuffer to the
OS for processing now get what they expect. In addition for applications
that want to do the right thing, they now have less work to do.